0504-AOP

+++ src/main/java/com/karina/app/board/notice/NoticeController.java
@@ -13,11 +13,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.ModelAndView;
 
 import com.karina.app.board.BoardDTO;
 import com.karina.app.file.FileDTO;
+import com.karina.app.member.MemberDTO;
 import com.karina.app.pager.Pager;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Controller
@@ -95,16 +98,33 @@ public class NoticeController {
 		return "board/update";
 	}
 	@PostMapping("update")
-	public String update(NoticeDTO noticeDTO,@RequestParam("attach") MultipartFile[]attach)throws Exception{
+	public ModelAndView update(NoticeDTO noticeDTO,@RequestParam("attach") MultipartFile[]attach,Model model)throws Exception{
 		int result=noticeService.update(noticeDTO,attach);
-		return "redirect:./list";
+		ModelAndView mv= new ModelAndView();
+		mv.setViewName("redirect:./list");
+		mv.addObject("update",noticeDTO);
+		return mv;
 		
 	}
 	@PostMapping("delete")
-	public String delete(NoticeDTO noticeDTO) throws Exception{
+	public String delete(NoticeDTO noticeDTO,HttpSession session,Model model) throws Exception{
+		MemberDTO memberDTO=(MemberDTO)session.getAttribute("member");
+		
+		BoardDTO boardDTO = noticeService.detail(noticeDTO);
+		if(boardDTO.getBoardWriter().equals(memberDTO.getUsername()))
+		{
 		int result= noticeService.delete(noticeDTO);
 		return "redirect:./list";
-	}
+		}
+		else {
+			model.addAttribute("result","작성자가 아닙니다");
+			model.addAttribute("url", "./list");
+			return "commons.result";
+			
+		}
+			
+		
+	}}
 
 
-}
+
